functioning communication between producer, consumer, and listener
