+++ src/queue/consumers/puppeteer/index.ts
@@ -17,24 +17,6 @@ export const setupPuppeteer = async (initVals: {
     );
   }
 
-  // Determine tor/proxy for browser
-  //if (isTor) {
-  //let ports = process.env.TOR_PORTS.split(",");
-  //let portIndex = getRandom(0, ports.length - 1)();
-  //let port = isProduction ? ports[portIndex] : "9050";
-  //proxy = `socks5://127.0.0.1:` + port;
-  //}
-
-  //if (isProxy) {
-  //let proxies = await getProxies();
-  //let proxyIndex = getRandom(0, proxies.length - 1)();
-  //proxy = proxies[proxyIndex];
-  //}
-
-  //if (isProxy || isTor) {
-  //args.push(`--proxy-server=${proxy}`);
-  //}
-
   ///// Initialize Browser
   const browser = await puppeteer.launch({
     headless: process.env.NODE_ENV === "production",
@@ -46,21 +28,6 @@ export const setupPuppeteer = async (initVals: {
   browser.on("disconnected", () => {
     console.log("Browser was disconnected.");
   });
-
-  /////// Page setup
-  //const page = (await browser.pages())[0];
-  //if (isTor) {
-  //await page.goto("https://check.torproject.org/");
-  //const isUsingTor = await page.$eval("body", (el) =>
-  //el.innerHTML.includes(
-  //"Congratulations. This browser is configured to use Tor"
-  //)
-  //);
-  //if (!isUsingTor) {
-  //logger.error(`Browser is not using Tor. Exiting...`);
-  //return await browser.close();
-  //}
-  //}
 
   console.log(
     `Configured through site ${isTor || isProxy ? proxy : "locally."}`
