+++ src/queue/index.ts
@@ -8,6 +8,25 @@ import { askQuestion } from "../util";
 // Import different jobs
 import { house, senate } from "./jobs";
 
+export interface Committee {
+  title: string;
+  link: string;
+  date?: Date;
+  time?: Date;
+  text?: string;
+  location?: string;
+}
+
+interface metaType {
+  collection: "houseCommittee" | "senateCommittee";
+  name: string;
+}
+
+export interface jobType {
+  data: Committee[];
+  meta: metaType;
+}
+
 export const setupQueue = async () => {
   // Create an instance of the Bull queue: https://github.com/OptimalBits/bull
   try {
